Spit all versions to namespaces
Each version of the stencil computation is now in its own namespace with
an initfunc_t member that is then selected in main.cu dynamically. This
allows us to produce a single binary for all versions, and makes it
easier to time the code/run benchmarks, as well as making the codebase
cleaner overall.