Moved blockdim calculation outside of omp parallel
There is no need to repeat this portion of the code for every GPU
dev_id. However, it should be noted that there is an implicit
requirement that all GPUs have the same memory and SM count, as a hybrid
system will overcomplicate the code. Perhaps do a check and throw a
warning?