More improvements and bug fixes.
Seems to work for domain sizes that we will use to collect performance data. For some reason, does not work for domain sizes that are not perfectly divisible by 256. Might be because of the how the grid_dim_x variable is calculated at the start of the kernel (Do we need to use tile_size or the remainder to calculate this?)