Removed boundary_buffer pointer addition
I am not fuly sure what what this boundary_buffer does, but it seems
to be only caching east and west halow? It seems like north and south
are cached normally and used in the computation as is. It is possible
that PERKS does not write to north and south halos (code is hard to
understand), which is good, as the boundary stream is the one that
needs to write there. I assume it's not all cached because it has
no writes. Anyways, we'll see